Remove unused whitenoise package dependency
This library is supposed to be used in combination with Django as a
middleware in order to have Django serve static files in production.

It doesn't seem like it has ever been set up or is even being used, and
we have already made plans to use nginx for serving said files, so
removal seems to be the best option.