Add multi-db routing
This commit adds django-multidb-router to the dependency graph and makes
the necessary changes so that it's enabled in stage and production.

This change makes it so that any write operations go to the main
read/write database endpoint and any read operations go to a replica
instead, this in turn means better performance and availability.